Fixed some stuff in the API

+++ test_project/UNTITLED/nirvana_stage.bas
@@ -66,8 +66,16 @@ NIRVANA_WINDOW_HEIGHT = 200
 
 Dim Nirvana_Stage_Sprites[9] As Nirvana_Sprite
 Dim Nirvana_Stage_Shape_Points[1] As Nirvana_Vector2D
-Dim Nirvana_Stage_Shapes[7] As Nirvana_Shape
+Dim Nirvana_Stage_Shapes[6] As Nirvana_Shape
 Dim Nirvana_Stage_Layers[5] As Nirvana_Layer
+
+Dim Nirvana_NULL_Shape As Nirvana_Shape
+Nirvana_NULL_Shape.Name$ = ""
+Nirvana_NULL_Shape.Sprite_ID = -1
+Nirvana_NULL_Shape.ShapeType = -1
+Nirvana_NULL_Shape.ShapeData_Matrix = -1
+
+ArrayFill(Nirvana_Stage_Shapes, Nirvana_NULL_Shape)
 
 For i = 0 To 4
 	Nirvana_Stage_Layers[i].Ref_Canvas = -1
@@ -82,7 +90,7 @@ For i = 0 To 8
 	Nirvana_Stage_Sprites[i].Sprite_ID = -1
 Next
 
-For i = 0 To 6
+For i = 0 To 5
 	Nirvana_Stage_Shapes[i].Sprite_ID = -1
 Next
 
@@ -105,12 +113,23 @@ Sub Nirvana_ClearStage()
 
 	Nirvana_ActiveStage.Stage_Offset.X = 0
 	Nirvana_ActiveStage.Stage_Offset.Y = 0
+
+	Num_Shapes = ArraySize(Nirvana_Stage_Shapes, 1)
+	For n = 0 To Num_Shapes-1
+		If MatrixExists(Nirvana_Stage_Shapes[n].ShapeData_Matrix) Then
+			DeleteMatrix(Nirvana_Stage_Shapes[n].ShapeData_Matrix)
+		End If
+		Nirvana_Stage_Shapes[n].ShapeData_Matrix = -1
+	Next
+
 	For i = 0 To Nirvana_ActiveStage.Layer_Count-1
 
 		CloseCanvas(Nirvana_Stage_Layers[i].Ref_Canvas)
 		Nirvana_Stage_Layers[i].Ref_Canvas = -1
 
-		DeleteStack_N(Nirvana_Stage_Layers[i].Layer_TS_Stack)
+		If Stack_N_Exists(Nirvana_Stage_Layers[i].Layer_TS_Stack) Then
+			DeleteStack_N(Nirvana_Stage_Layers[i].Layer_TS_Stack)
+		End If
 		Nirvana_Stage_Layers[i].Layer_TS_Stack = -1
 
 		If ImageExists(Nirvana_Stage_Layers[i].Bkg.Image_ID) Then
@@ -223,7 +242,7 @@ Function Nirvana_Stage_0(vp_x, vp_y, vp_w, vp_h) As Nirvana_Stage
 	SetSpriteCanvasRenderPriority(Nirvana_Stage_Layers[2].Ref_Canvas, SPRITE_PRIORITY_GREATEST_Y, SPRITE_ORDER_DESCENDING)
 	Canvas(Nirvana_Stage_Layers[2].Ref_Canvas)
 	Nirvana_Stage_Layers[2].Layer_Sprite_Count = 8
-	Nirvana_Stage_Layers[2].Layer_Shape_Count = 7
+	Nirvana_Stage_Layers[2].Layer_Shape_Count = 5
 
 	'------- SPRITE ("g1") -------
 	Nirvana_Stage_Sprites[0] = Nirvana_CreateSprite("grz", "g1")
@@ -400,49 +419,29 @@ Function Nirvana_Stage_0(vp_x, vp_y, vp_w, vp_h) As Nirvana_Stage
 	'Base Settings
 	'Shape
 	Nirvana_Stage_Shapes[0].ShapeType = SPRITE_SHAPE_POLYGON
-	Dim stage_shape_x_0[8]
-	Dim stage_shape_y_0[8]
-	Nirvana_Stage_Shapes[0].ShapeData_Matrix = DimMatrix(2, 8)
+	Dim stage_shape_x_0[4]
+	Dim stage_shape_y_0[4]
+	Nirvana_Stage_Shapes[0].ShapeData_Matrix = DimMatrix(2, 4)
 
-	stage_shape_x_0[0] = 1026
-	stage_shape_y_0[0] = 177
+	stage_shape_x_0[0] = 671
+	stage_shape_y_0[0] = 242
 	SetMatrixValue(Nirvana_Stage_Shapes[0].ShapeData_Matrix, 0, 0, stage_shape_x_0[0])
 	SetMatrixValue(Nirvana_Stage_Shapes[0].ShapeData_Matrix, 1, 0, stage_shape_y_0[0])
 
-	stage_shape_x_0[1] = 885
-	stage_shape_y_0[1] = 179
+	stage_shape_x_0[1] = 548
+	stage_shape_y_0[1] = 340
 	SetMatrixValue(Nirvana_Stage_Shapes[0].ShapeData_Matrix, 0, 1, stage_shape_x_0[1])
 	SetMatrixValue(Nirvana_Stage_Shapes[0].ShapeData_Matrix, 1, 1, stage_shape_y_0[1])
 
-	stage_shape_x_0[2] = 919
-	stage_shape_y_0[2] = 310
+	stage_shape_x_0[2] = 683
+	stage_shape_y_0[2] = 450
 	SetMatrixValue(Nirvana_Stage_Shapes[0].ShapeData_Matrix, 0, 2, stage_shape_x_0[2])
 	SetMatrixValue(Nirvana_Stage_Shapes[0].ShapeData_Matrix, 1, 2, stage_shape_y_0[2])
 
-	stage_shape_x_0[3] = 941
-	stage_shape_y_0[3] = 301
+	stage_shape_x_0[3] = 792
+	stage_shape_y_0[3] = 345
 	SetMatrixValue(Nirvana_Stage_Shapes[0].ShapeData_Matrix, 0, 3, stage_shape_x_0[3])
 	SetMatrixValue(Nirvana_Stage_Shapes[0].ShapeData_Matrix, 1, 3, stage_shape_y_0[3])
-
-	stage_shape_x_0[4] = 931
-	stage_shape_y_0[4] = 200
-	SetMatrixValue(Nirvana_Stage_Shapes[0].ShapeData_Matrix, 0, 4, stage_shape_x_0[4])
-	SetMatrixValue(Nirvana_Stage_Shapes[0].ShapeData_Matrix, 1, 4, stage_shape_y_0[4])
-
-	stage_shape_x_0[5] = 981
-	stage_shape_y_0[5] = 209
-	SetMatrixValue(Nirvana_Stage_Shapes[0].ShapeData_Matrix, 0, 5, stage_shape_x_0[5])
-	SetMatrixValue(Nirvana_Stage_Shapes[0].ShapeData_Matrix, 1, 5, stage_shape_y_0[5])
-
-	stage_shape_x_0[6] = 938
-	stage_shape_y_0[6] = 340
-	SetMatrixValue(Nirvana_Stage_Shapes[0].ShapeData_Matrix, 0, 6, stage_shape_x_0[6])
-	SetMatrixValue(Nirvana_Stage_Shapes[0].ShapeData_Matrix, 1, 6, stage_shape_y_0[6])
-
-	stage_shape_x_0[7] = 976
-	stage_shape_y_0[7] = 335
-	SetMatrixValue(Nirvana_Stage_Shapes[0].ShapeData_Matrix, 0, 7, stage_shape_x_0[7])
-	SetMatrixValue(Nirvana_Stage_Shapes[0].ShapeData_Matrix, 1, 7, stage_shape_y_0[7])
 
 
 
@@ -555,107 +554,6 @@ Function Nirvana_Stage_0(vp_x, vp_y, vp_w, vp_h) As Nirvana_Stage
 
 
 
-	'------- SHAPE ("POLYGON_2") -------
-	Nirvana_Stage_Shapes[5].Name$ = "POLYGON_2"
-	Nirvana_Stage_Shapes[5].Sprite_ID = -1
-
-	'Base Settings
-	'Shape
-	Nirvana_Stage_Shapes[5].ShapeType = SPRITE_SHAPE_POLYGON
-	Dim stage_shape_x_5[1]
-	Dim stage_shape_y_5[1]
-	Nirvana_Stage_Shapes[5].ShapeData_Matrix = DimMatrix(2, 1)
-
-	stage_shape_x_5[0] = 126
-	stage_shape_y_5[0] = 294
-	SetMatrixValue(Nirvana_Stage_Shapes[5].ShapeData_Matrix, 0, 0, stage_shape_x_5[0])
-	SetMatrixValue(Nirvana_Stage_Shapes[5].ShapeData_Matrix, 1, 0, stage_shape_y_5[0])
-
-
-
-	'------- SHAPE ("POLYGON_3") -------
-	Nirvana_Stage_Shapes[6].Name$ = "POLYGON_3"
-	Nirvana_Stage_Shapes[6].Sprite_ID = -1
-
-	'Base Settings
-	'Shape
-	Nirvana_Stage_Shapes[6].ShapeType = SPRITE_SHAPE_POLYGON
-	Dim stage_shape_x_6[14]
-	Dim stage_shape_y_6[14]
-	Nirvana_Stage_Shapes[6].ShapeData_Matrix = DimMatrix(2, 14)
-
-	stage_shape_x_6[0] = 1304
-	stage_shape_y_6[0] = 307
-	SetMatrixValue(Nirvana_Stage_Shapes[6].ShapeData_Matrix, 0, 0, stage_shape_x_6[0])
-	SetMatrixValue(Nirvana_Stage_Shapes[6].ShapeData_Matrix, 1, 0, stage_shape_y_6[0])
-
-	stage_shape_x_6[1] = 1334
-	stage_shape_y_6[1] = 354
-	SetMatrixValue(Nirvana_Stage_Shapes[6].ShapeData_Matrix, 0, 1, stage_shape_x_6[1])
-	SetMatrixValue(Nirvana_Stage_Shapes[6].ShapeData_Matrix, 1, 1, stage_shape_y_6[1])
-
-	stage_shape_x_6[2] = 1279
-	stage_shape_y_6[2] = 395
-	SetMatrixValue(Nirvana_Stage_Shapes[6].ShapeData_Matrix, 0, 2, stage_shape_x_6[2])
-	SetMatrixValue(Nirvana_Stage_Shapes[6].ShapeData_Matrix, 1, 2, stage_shape_y_6[2])
-
-	stage_shape_x_6[3] = 1183
-	stage_shape_y_6[3] = 388
-	SetMatrixValue(Nirvana_Stage_Shapes[6].ShapeData_Matrix, 0, 3, stage_shape_x_6[3])
-	SetMatrixValue(Nirvana_Stage_Shapes[6].ShapeData_Matrix, 1, 3, stage_shape_y_6[3])
-
-	stage_shape_x_6[4] = 1119
-	stage_shape_y_6[4] = 321
-	SetMatrixValue(Nirvana_Stage_Shapes[6].ShapeData_Matrix, 0, 4, stage_shape_x_6[4])
-	SetMatrixValue(Nirvana_Stage_Shapes[6].ShapeData_Matrix, 1, 4, stage_shape_y_6[4])
-
-	stage_shape_x_6[5] = 1227
-	stage_shape_y_6[5] = 244
-	SetMatrixValue(Nirvana_Stage_Shapes[6].ShapeData_Matrix, 0, 5, stage_shape_x_6[5])
-	SetMatrixValue(Nirvana_Stage_Shapes[6].ShapeData_Matrix, 1, 5, stage_shape_y_6[5])
-
-	stage_shape_x_6[6] = 1361
-	stage_shape_y_6[6] = 248
-	SetMatrixValue(Nirvana_Stage_Shapes[6].ShapeData_Matrix, 0, 6, stage_shape_x_6[6])
-	SetMatrixValue(Nirvana_Stage_Shapes[6].ShapeData_Matrix, 1, 6, stage_shape_y_6[6])
-
-	stage_shape_x_6[7] = 1375
-	stage_shape_y_6[7] = 397
-	SetMatrixValue(Nirvana_Stage_Shapes[6].ShapeData_Matrix, 0, 7, stage_shape_x_6[7])
-	SetMatrixValue(Nirvana_Stage_Shapes[6].ShapeData_Matrix, 1, 7, stage_shape_y_6[7])
-
-	stage_shape_x_6[8] = 1283
-	stage_shape_y_6[8] = 457
-	SetMatrixValue(Nirvana_Stage_Shapes[6].ShapeData_Matrix, 0, 8, stage_shape_x_6[8])
-	SetMatrixValue(Nirvana_Stage_Shapes[6].ShapeData_Matrix, 1, 8, stage_shape_y_6[8])
-
-	stage_shape_x_6[9] = 1123
-	stage_shape_y_6[9] = 443
-	SetMatrixValue(Nirvana_Stage_Shapes[6].ShapeData_Matrix, 0, 9, stage_shape_x_6[9])
-	SetMatrixValue(Nirvana_Stage_Shapes[6].ShapeData_Matrix, 1, 9, stage_shape_y_6[9])
-
-	stage_shape_x_6[10] = 1086
-	stage_shape_y_6[10] = 352
-	SetMatrixValue(Nirvana_Stage_Shapes[6].ShapeData_Matrix, 0, 10, stage_shape_x_6[10])
-	SetMatrixValue(Nirvana_Stage_Shapes[6].ShapeData_Matrix, 1, 10, stage_shape_y_6[10])
-
-	stage_shape_x_6[11] = 1104
-	stage_shape_y_6[11] = 340
-	SetMatrixValue(Nirvana_Stage_Shapes[6].ShapeData_Matrix, 0, 11, stage_shape_x_6[11])
-	SetMatrixValue(Nirvana_Stage_Shapes[6].ShapeData_Matrix, 1, 11, stage_shape_y_6[11])
-
-	stage_shape_x_6[12] = 1166
-	stage_shape_y_6[12] = 422
-	SetMatrixValue(Nirvana_Stage_Shapes[6].ShapeData_Matrix, 0, 12, stage_shape_x_6[12])
-	SetMatrixValue(Nirvana_Stage_Shapes[6].ShapeData_Matrix, 1, 12, stage_shape_y_6[12])
-
-	stage_shape_x_6[13] = 1272
-	stage_shape_y_6[13] = 427
-	SetMatrixValue(Nirvana_Stage_Shapes[6].ShapeData_Matrix, 0, 13, stage_shape_x_6[13])
-	SetMatrixValue(Nirvana_Stage_Shapes[6].ShapeData_Matrix, 1, 13, stage_shape_y_6[13])
-
-
-
 	'------- TILE SPRITES  -------
 	Cut_Start_Index = 0
 	Nirvana_Stage_Layers[2].Layer_TS_Stack = CreateStack_N()
@@ -737,7 +635,7 @@ Function Nirvana_Stage_0(vp_x, vp_y, vp_w, vp_h) As Nirvana_Stage
 	SetSpriteCanvasRenderPriority(Nirvana_Stage_Layers[4].Ref_Canvas, SPRITE_PRIORITY_NONE, SPRITE_ORDER_ASCENDING)
 	Canvas(Nirvana_Stage_Layers[4].Ref_Canvas)
 	Nirvana_Stage_Layers[4].Layer_Sprite_Count = 1
-	Nirvana_Stage_Layers[4].Layer_Shape_Count = 0
+	Nirvana_Stage_Layers[4].Layer_Shape_Count = 1
 
 	'------- SPRITE ("guy1") -------
 	Nirvana_Stage_Sprites[8] = Nirvana_CreateSprite("gren_guy", "guy1")
@@ -779,6 +677,27 @@ Function Nirvana_Stage_0(vp_x, vp_y, vp_w, vp_h) As Nirvana_Stage
 	'Render Settings
 	SetSpriteAlpha(Nirvana_Stage_Sprites[8].Sprite_ID, 180)
 	SetSpriteVisible(Nirvana_Stage_Sprites[8].Sprite_ID, TRUE)
+
+	'------- SHAPE ("CIRCLE_1") -------
+	Nirvana_Stage_Shapes[5].Name$ = "CIRCLE_1"
+	Nirvana_Stage_Shapes[5].Sprite_ID = CreateSprite(-1, 1, 1)
+
+	'Base Settings
+	SetSpriteType(Nirvana_Stage_Shapes[5].Sprite_ID, SPRITE_TYPE_STATIC)
+	SetSpriteSolid(Nirvana_Stage_Shapes[5].Sprite_ID, TRUE)
+	'Shape
+	Nirvana_Stage_Shapes[5].ShapeType = SPRITE_SHAPE_CIRCLE
+	SetSpriteShape(Nirvana_Stage_Shapes[5].Sprite_ID, Nirvana_Stage_Shapes[5].ShapeType)
+	Nirvana_Stage_Shapes[5].ShapeData_Matrix = DimMatrix(3, 1)
+	SetMatrixValue(Nirvana_Stage_Shapes[5].ShapeData_Matrix, 0, 0, 135)
+	SetMatrixValue(Nirvana_Stage_Shapes[5].ShapeData_Matrix, 1, 0, 87)
+	SetMatrixValue(Nirvana_Stage_Shapes[5].ShapeData_Matrix, 2, 0, 63)
+	SetSpriteShapeOffset(Nirvana_Stage_Shapes[5].Sprite_ID, 135, 87)
+	SetSpriteRadius(Nirvana_Stage_Shapes[5].Sprite_ID, 63)
+
+	'Transform
+	SetSpritePosition(Nirvana_Stage_Shapes[5].Sprite_ID,  0, 0) 'Set to (0,0) so that offset will be world position
+	SetSpriteVisible(Nirvana_Stage_Shapes[5].Sprite_ID, FALSE) 'Image is -1 so it wouldn't render anyway
 
 	'------- TILE SPRITES  -------
 	Cut_Start_Index = 4
